fix issue by replacing type intent
